+++ packages/desktop_multi_window/lib/src/window_controller_impl.dart
@@ -1,3 +1,4 @@
+import 'dart:io';
 import 'dart:ui';
 
 import 'package:flutter/services.dart';
@@ -6,7 +7,7 @@ import 'channels.dart';
 import 'window_controller.dart';
 
 class WindowControllerMainImpl extends WindowController {
-  final MethodChannel _channel = miltiWindowChannel;
+  final MethodChannel _channel = multiWindowChannel;
 
   // the id of this window
   final int _id;
@@ -53,6 +54,20 @@ class WindowControllerMainImpl extends WindowController {
       'windowId': _id,
       'title': title,
     });
+  }
+
+  @override
+  Future<void> resizable(bool resizable) {
+    if (Platform.isMacOS) {
+      return _channel.invokeMethod('resizable', <String, dynamic>{
+        'windowId': _id,
+        'resizable': resizable,
+      });
+    } else {
+      throw MissingPluginException(
+        'This functionality is only available on macOS',
+      );
+    }
   }
 
   @override