+++ packages/desktop_multi_window/macos/Classes/FlutterMultiWindowPlugin.swift
@@ -67,6 +67,12 @@ public class FlutterMultiWindowPlugin: NSObject, FlutterPlugin {
       let title = arguments["title"] as! String
       MultiWindowManager.shared.setTitle(windowId: windowId, title: title)
       result(nil)
+	case "resizable":
+	  let arguments = call.arguments as! [String: Any?]
+	  let windowId = arguments["windowId"] as! Int64
+	  let resizable = arguments["resizable"] as! Bool
+	  MultiWindowManager.shared.resizable(windowId: windowId, resizable: resizable)
+	  result(nil)
     case "setFrameAutosaveName":
       let arguments = call.arguments as! [String: Any?]
       let windowId = arguments["windowId"] as! Int64