+++ packages/mixin_logger/lib/src/write_log_to_file_web.dart
@@ -4,8 +4,8 @@ void writeLog(String log) {
   // do nothing.
 }
 
-void initLogger(
+Future<void> initLogger(
   String logDir,
   int maxFileCount,
   int maxFileLength,
-) {}
+) async {}