+++ packages/ogg_opus_player/lib/src/player_plugin_impl.dart
@@ -22,47 +22,76 @@ PlayerState _convertFromRawValue(int state) {
   }
 }
 
-class OggOpusPlayerPluginImpl extends OggOpusPlayer {
-  static const MethodChannel _channel = MethodChannel('ogg_opus_player');
+const MethodChannel _channel = MethodChannel('ogg_opus_player');
 
-  static var _inited = false;
+final Map<int, OggOpusPlayerPluginImpl> _players = {};
+final Map<int, OggOpusRecorderPluginImpl> _recorders = {};
 
-  static final Map<int, OggOpusPlayerPluginImpl> _players = {};
+var _initialized = false;
 
-  static Future<dynamic> _handleMethodCall(MethodCall call) async {
-    switch (call.method) {
-      case "onPlayerStateChanged":
-        final state = call.arguments['state'] as int;
-        final position = call.arguments['position'] as double;
-        final playerId = call.arguments['playerId'] as int;
-        final updateTime = call.arguments['updateTime'] as int;
-        final player = _players[playerId];
-        if (player == null) {
-          return;
-        }
-        player._playerState.value = _convertFromRawValue(state);
-        player._lastUpdateTimeStampe = updateTime;
-        player._position = position;
-        break;
-      default:
-        break;
-    }
+void _initChannelIfNeeded() {
+  if (_initialized) {
+    return;
   }
-
-  static void _initChannelIfNeeded() {
-    if (_inited) {
-      return;
+  _initialized = true;
+  _channel.setMethodCallHandler((call) async {
+    try {
+      return await _handleMethodCall(call);
+    } catch (e) {
+      debugPrint("_handleMethodCall: $e");
     }
-    _inited = true;
-    _channel.setMethodCallHandler((call) async {
-      try {
-        return await _handleMethodCall(call);
-      } catch (e) {
-        debugPrint("_handleMethodCall: $e");
+  });
+}
+
+Future<dynamic> _handleMethodCall(MethodCall call) async {
+  switch (call.method) {
+    case "onPlayerStateChanged":
+      final state = call.arguments['state'] as int;
+      final position = call.arguments['position'] as double;
+      final playerId = call.arguments['playerId'] as int;
+      final updateTime = call.arguments['updateTime'] as int;
+      final player = _players[playerId];
+      if (player == null) {
+        return;
       }
-    });
+      player._playerState.value = _convertFromRawValue(state);
+      player._lastUpdateTimeStamp = updateTime;
+      player._position = position;
+      break;
+    case "onRecorderCanceled":
+      final recorderId = call.arguments['recorderId'] as int;
+      final recorder = _recorders[recorderId];
+      final reason = call.arguments['reason'] as int;
+      if (recorder == null) {
+        return;
+      }
+      recorder.onCanceled(reason);
+      break;
+    case "onRecorderStartFailed":
+      final recorderId = call.arguments['recorderId'] as int;
+      final recorder = _recorders[recorderId];
+      if (recorder == null) {
+        return;
+      }
+      final reason = call.arguments['error'] as String;
+      debugPrint('onRecorderStartFailed: $reason');
+      break;
+    case "onRecorderFinished":
+      final recorderId = call.arguments['recorderId'] as int;
+      final recorder = _recorders[recorderId];
+      if (recorder == null) {
+        return;
+      }
+      final duration = call.arguments['duration'] as int;
+      final waveform = (call.arguments['waveform'] as List).cast<int>();
+      recorder.onFinished(duration, waveform);
+      break;
+    default:
+      break;
   }
+}
 
+class OggOpusPlayerPluginImpl extends OggOpusPlayer {
   OggOpusPlayerPluginImpl(this._path) : super.create() {
     _initChannelIfNeeded();
     assert(() {
@@ -99,17 +128,17 @@ class OggOpusPlayerPluginImpl extends OggOpusPlayer {
   double _position = 0.0;
 
   // [_position] updated timestamp, in milliseconds.
-  int _lastUpdateTimeStampe = -1;
+  int _lastUpdateTimeStamp = -1;
 
   @override
   double get currentPosition {
-    if (_lastUpdateTimeStampe == -1) {
+    if (_lastUpdateTimeStamp == -1) {
       return 0;
     }
     if (state.value != PlayerState.playing) {
       return _position;
     }
-    final offset = SystemClock.uptime().inMilliseconds - _lastUpdateTimeStampe;
+    final offset = SystemClock.uptime().inMilliseconds - _lastUpdateTimeStamp;
     assert(offset >= 0);
     if (offset < 0) {
       return _position;
@@ -143,5 +172,74 @@ class OggOpusPlayerPluginImpl extends OggOpusPlayer {
   @override
   void dispose() {
     _channel.invokeMethod("stop", _playerId);
+  }
+}
+
+class OggOpusRecorderPluginImpl extends OggOpusRecorder {
+  OggOpusRecorderPluginImpl(this._path) : super.create() {
+    _initChannelIfNeeded();
+    scheduleMicrotask(() async {
+      try {
+        _id = await _channel.invokeMethod('createRecorder', _path);
+        _recorders[_id] = this;
+      } catch (e) {
+        debugPrint('create recorder failed. error: $e');
+      }
+      _createCompleter.complete();
+    });
+  }
+
+  final String _path;
+  int _id = -1;
+
+  double? _duration;
+  List<int>? _waveformData;
+
+  final _createCompleter = Completer<void>();
+
+  final _stopCompleter = Completer<void>();
+
+  @override
+  Future<void> start() async {
+    await _createCompleter.future;
+    if (_id <= 0) {
+      return;
+    }
+    await _channel.invokeMethod('startRecord', _id);
+  }
+
+  @override
+  Future<void> stop() async {
+    await _createCompleter.future;
+    if (_id <= 0) {
+      return;
+    }
+    await _channel.invokeMethod('stopRecord', _id);
+    await _stopCompleter.future;
+  }
+
+  @override
+  void dispose() {
+    _channel.invokeMethod("destroyRecorder", _id);
+  }
+
+  @override
+  Future<double> duration() async {
+    return _duration ?? 0.0;
+  }
+
+  @override
+  Future<List<int>> getWaveformData() async {
+    return _waveformData ?? [];
+  }
+
+  void onCanceled(int reason) {
+    _stopCompleter.complete();
+  }
+
+  void onFinished(int duration, List<int> waveform) {
+    _duration = duration / 1000;
+    _waveformData = waveform;
+    _stopCompleter.complete();
   }
 }