+++ packages/ogg_opus_player/CHANGELOG.md
@@ -1,3 +1,11 @@
+## 0.4.1
+
+[Linux] replace opus static libraries to release version.
+
+## 0.4.0
+
+Support recording voice.
+
 ## 0.3.2
 
 Fix wrong playback speed on some devices.
@@ -12,11 +20,12 @@ support windows and linux.
 
 ## 0.2.0
 
-suppport macos arm64
+support macOS arm64
 
 ## 0.1.0
 
 break change: upgrade iOS min version from 10.0 to 12.0
+
 ## 0.0.1
 
 add iOS/ macOS support.